pyproject.toml
[tool.poetry]
name = "generate_bar_code_telegram_bot"
version = "0.1.0"
description = "Generate Barcode within telegram bot ( python language, server side)"
authors = ["Serhii Hrekov <themebeat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^13.7"
python-barcode = "^0.15.1"
pillow = "10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
